requirements_dev.txt
# Django
Django==1.9.8

# Django Rest Framework
djangorestframework==3.5.1
Markdown==2.6.7
django-filter==0.15.3
